Update VOLTTRON-Central-Demo.rst
Slight refinement in the example.  We DO want the instance name of the 2nd instance to be different than the 1st instance name (otherwise it wont show up at all as a platform in VOLTTRON Central).  Hence need to show in the example that we set the 2nd instance name to 'volttron2'.